Enemies

+++ src/Mapa.cpp
@@ -1,12 +1,5 @@
 #include "Mapa.h"
-#include "raymath.h"
-#include "Mapa.h"
-#include "raymath.h"
-#include "raylib.h"
-#include <json/json.h>
-#include <glog/logging.h>
 
-// Resto del código...
 
 Mapa::Mapa() {
     // Añadir texturas
@@ -15,16 +8,19 @@ Mapa::Mapa() {
     UnloadImage(image);
 }
 
-void Mapa::DrawTile(int pos_x, int pos_y, int tile_x, int tile_y, texture_asset texture) {
-    Rectangle sourceRec = { (float)TILE_SIZE * tile_x, (float)TILE_SIZE * tile_y, (float)TILE_SIZE, (float)TILE_SIZE };
-    Rectangle destRec = { (float)pos_x, (float)pos_y, (float)TILE_SIZE, (float)TILE_SIZE };
-    Vector2 origin = { 0, 0 };
+// Método para dibujar un tile
+void Mapa::DrawTile( int pos_x, int pos_y, int tile_x, int tile_y, texture_asset texture){
+    Rectangle sourceRec = {(float)TILE_SIZE * tile_x, (float)TILE_SIZE * tile_y, (float)TILE_SIZE, (float)TILE_SIZE };
+    Rectangle destRec = {(float)pos_x, (float)pos_y, (float)TILE_SIZE, (float)TILE_SIZE};
+    Vector2 origin = {0, 0};
     DrawTexturePro(textures[texture], sourceRec, destRec, origin, 0.0f, WHITE);
+
 }
 
+// Método para dibujar un sólo mapa
 void Mapa::DrawMap(int matriz[][MAP_HEIGHT], int tileSetSize, texture_asset texture) {
-    for (int y = 0; y < MAP_HEIGHT; y++) {
-        for (int x = 0; x < MAP_WIDTH; x++) {
+    for (int y = 0; y < MAP_HEIGHT; y++){
+        for (int x = 0; x < MAP_WIDTH; x++){
             int tileID = matriz[x][y];
             int tile_x = (tileID - 1) % tileSetSize;
             int tile_y = (tileID - 1) / tileSetSize;
@@ -34,8 +30,6 @@ void Mapa::DrawMap(int matriz[][MAP_HEIGHT], int tileSetSize, texture_asset text
     }
 }
 
-// Método para dibujar un mapa en una posición específica con un área circular
-// Método para dibujar un mapa en una posición específica con un área circular
 void Mapa::DrawMapAtPosition(int matriz[][MAP_HEIGHT], int tileSetSize, texture_asset texture, Vector2 center, float innerRadius, float outerRadius) {
     // Calcular el radio máximo permitido para dibujar
     float maxRadius = outerRadius - TILE_SIZE / 2;
@@ -47,7 +41,7 @@ void Mapa::DrawMapAtPosition(int matriz[][MAP_HEIGHT], int tileSetSize, texture_
             int tile_y = y * TILE_SIZE;
 
             // Calcular la distancia desde el centro a la posición actual del tile
-            float distance = Vector2Distance(center, { tile_x + TILE_SIZE / 2, tile_y + TILE_SIZE / 2 });
+            float distance = Vector2Distance(center, { static_cast<float>(tile_x + TILE_SIZE / 2), static_cast<float>(tile_y + TILE_SIZE / 2) });
 
             // Verificar si la distancia está dentro del rango del radio interior y exterior
             if (distance >= innerRadius && distance <= maxRadius) {
@@ -64,4 +58,3 @@ void Mapa::DrawMapAtPosition(int matriz[][MAP_HEIGHT], int tileSetSize, texture_
         }
     }
 }
-
